Changed MMVAE to be example and specified setup.py versions with ml-python/nightly versions
