+++ d_mmvae/models/__init__.py
@@ -1,3 +1,2 @@
 from d_mmvae.models.Models import VAE, Expert
-
 __all__ = ['Expert', 'VAE']
