+++ d_mmvae/trainers/__init__.py
@@ -1,3 +1,3 @@
-from d_mmvae.trainers.MMVAE import MMVAETrainer
+from d_mmvae.trainers.ExampleTrainer import ExampleTrainer
 from d_mmvae.trainers.trainer import BaseTrainer
-__all__ = ['MMVAETrainer', 'BaseTrainer']
+__all__ = ['ExampleTrainer', 'BaseTrainer']
